+++ src/Symfony/Component/DomCrawler/Field/ChoiceFormField.php
@@ -153,11 +153,11 @@ class ChoiceFormField extends FormField
      *
      * This method should only be used internally.
      *
-     * @param \DOMElement $node
+     * @param \DOMNode $node A \DOMNode
      *
      * @throws \LogicException When choice provided is not multiple nor radio
      */
-    public function addChoice(\DOMElement $node)
+    public function addChoice(\DOMNode $node)
     {
         if (!$this->multiple && 'radio' !== $this->type) {
             throw new \LogicException(sprintf('Unable to add a choice for "%s" as it is not multiple or is not a radio button.', $this->name));
@@ -251,11 +251,11 @@ class ChoiceFormField extends FormField
     /**
      * Returns option value with associated disabled flag.
      *
-     * @param \DOMElement $node
+     * @param \DOMNode $node
      *
      * @return array
      */
-    private function buildOptionValue(\DOMElement $node)
+    private function buildOptionValue($node)
     {
         $option = array();